refactor(dh): move function to own file
